added base isr, wordisr, documentisr, andisr and indexreader, will test tomorrow

+++ include/fb/indexEntry.hpp
@@ -22,7 +22,7 @@ namespace fb {
  * mem = add_num(mem, 1623);
  * mem = add_num(mem, 610516583);
  */
-inline char* add_word_post( char* curr, size_t num) 
+inline char* add_word_post( char* curr, uint32_t num) 
    {
       if ( num <= fbImpl::oneBytePostMaxVal ) 
          {
@@ -52,7 +52,7 @@ inline char* add_word_sentinel( char* curr )
  * mem = read_number(mem, value);
  * cout << value << endl;
  */
-inline char* read_word_post( char* curr, uint64_t &num) 
+inline char* read_word_post( char* curr, uint32_t &num) 
    {
       switch ( ( ( fbImpl::oneBytePost * ) curr )->size ) 
          {
@@ -75,7 +75,7 @@ inline bool is_word_sentinel( char* curr )
    }
 
 // Adds document post. Returns the pointer to next address we should add to
-inline char* add_document_post( char* curr, size_t delta, uint32_t url_loc) {
+inline char* add_document_post( char* curr, uint32_t delta, uint32_t url_loc) {
    (* (uint32_t *) curr) = delta;
    curr += sizeof(delta);
    (* (uint32_t *) curr) = url_loc;
@@ -88,7 +88,7 @@ inline char* add_document_sentinel( char* curr )
    }
 
 // Reads document post. Returns the pointer to next address we should read from
-inline char* read_document_post( char* curr, size_t& delta, uint32_t& url_loc ) {
+inline char* read_document_post( char* curr, uint32_t& delta, uint32_t& url_loc ) {
    delta = (* (uint32_t *) curr);
    curr += sizeof(delta);
    url_loc = (* (uint32_t *) curr);
